Move network configuration to MediaBrowser.Common
