+++ Jellyfin.Api/Middleware/LanFilteringMiddleware.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using Jellyfin.Networking.Configuration;
 using MediaBrowser.Common.Extensions;
 using MediaBrowser.Common.Net;
 using MediaBrowser.Controller.Configuration;
