+++ tests/Jellyfin.Server.Tests/ParseNetworkTests.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Linq;
 using System.Net;
-using Jellyfin.Networking.Configuration;
 using Jellyfin.Networking.Manager;
 using Jellyfin.Server.Extensions;
 using MediaBrowser.Common.Configuration;
+using MediaBrowser.Common.Net;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.Extensions.Configuration;
